re-rexpot tacing modules
